tryToBeSmart = True
rather than the "incremental" approach of figuring out exactly
which attributes of which objects get changed, using a "monolithic"
approach that just re-runs the new expression. this circumvents
the issue of real vs. fake attributes for now. can revisit the
incremental approach later, if/when this becomes a performance
bottleneck.