Allow double-quoted strings.
After an hour of trying to figure out why my additions to prelude were
not parsing…it’s time we added this.

Required splitting out the type for expressions of base values and
base values themselves because the expressions need to keep track of
which quote character they used whereas the values themselves don’t care.