Remove failing tests.
I believe that these are actually fine to remove, based on the grammar
as described in the README.

There remain a few references to the [| ... |] form in the codebase --
deleteBoxes.little in the examples, the iBracks variable in UTils. However,
neither of those ever appear to be referenced in the code base, so I believe
that they are also safe to remove in a future cl.