Fix cursor positioning bug after code updates
Previously, whenever a code update occurred (e.g., via manipulation from
the GUI or from the run button), the cursor's position was reset to (0, 0).
The reason for this is that the model never actually got information
about the cursor's position until it was too late (after it was reset to
the top left); additionally, `Controller.msgFromAce` did not even track
the changes to the code box info, so that information was being thrown
away.

To fix this, this commit introduces a new "Ace Update" message that is
solely for receiving editor/document state updates and does nothing but
update the model to reflect this. Note that `Controller.msgFromAce` would
not satisfy this property because it also calls `upstateRun`, which
would re-run the entire file every single time a character changed if
used in this scenario.