Implement rudimentary saving and loading
For now, buffer-type saving is available (users can save and load to a
single buffer). Every time the code is updated, a `needsSave` flag in
the model is set to `True`, and every time the user presses the save
button (and the save successfully completes), the flag is set back to
false.

The buffer being used for this functionality is HTML5 localStorage.