Only allow save when code has changed
This commit makes it so that the editor only allows a save when the code
differs from what is saved in the buffer. This is more intuitive
behavior (before, every single update re-enabled the save button).

Additionally, to go along with this update, the code editor is now
initialized with whatever code was saved previously in the localStorage
buffer from the last editing session.