Add skeleton of filesystem
This commit adds all the necessary buttons for the new filesystem, as
well as updates the JavaScript API for handling files.

The new JavaScript API deals only with writing and requesting files; the
notion of "saving", "opening", etc. is something that only happens at
a higher level. Additionally, filenames are now required by the API.