Add deleting ability
For now, the "Delete" button simply resides in the "Open" dialog box,
but in the future there may be a file manager that is separate from the
"Open" window, allowing for the creation/deletion/opening of files from
a centralized location.

In order to handle deletion, the File Handling API has been updated to
include a "delete" command port, which sends an updated file index via
the "receiveFileIndex" subscription port after the file has been
deleted.