Add code importing
Users can now import a Little program from their computer to the
Sketch-n-Sketch editor.

To do so, the File Handling API has been updated.

Additionally, the mechanism for handling dialog boxes had to be slightly
reworked---now, instead of adding and removing HTML code on the fly, the
HTML for the dialog boxes is always present, and the CSS "display"
property simply toggles from "none" to "block" when the dialog is
needed. This is to avoid the input fields being reset every time the
view needs to update (like when an alert dialog box pops up over a
file dialog box).