Demonstrate faster loading time
With the Prelude stripped down immensely, it can be seen that
Sketch-n-Sketch loads much, much faster. Additionally, the stack
overflow error does not occur when more memory is not requested in
Chrome.

I strongly suspect that both of these setbacks are due to the Little
parser (and this commit provides some evidence for that). When the
Prelude is stripped down to almost nothing, the parser has very little
Little code to parse, and so the program can operate much more quickly.