Ostensibly correct moveDefinitionPat; UI finicky.
Before/After are now constructors: for patterns the given target position with before/after generated by UI interactions is internally translated into a simple PatternId which will be the location of the inserted pattern (before/after unnecessary). Before/after may or may not be removed from the UI layer in the future.

Still need to handle inserting and removing from EFun/ECase, as well as more complex patterns.