Add new file template categories
Templates can now be grouped into categories in the new file dialog box.
To change these categories, edit `ExamplesTemplate.elm`, then recompile
the examples.

Note that the order of the categories and the templates within the
category is preserved, so if alphabetizing is desired, manual sorting
based on the first element (the name) of the `makeExample` pair is
needed. This can be done by calling `List.sortBy fst`.