Consolidate reorder callee parameters and reorder list elements.
After the fact, I realized this wasn’t strictly necessary—Ravi’s code was correct, the crash was from Make Equal.

However, with a bit of massaging this version can now rearrange nested tuples (e.g. [a [b c]] to [a b [c]]) so I guess it’s a win.