parsing/unparsing implicit main expression
If the main expression is omitted, the parser inserts the
variable expression "main" (which assumes that main has been
defined at the top-level). Actually, it inserts the
expression "(let _IMPLICIT_MAIN '...' main)" to make it
easier for the unparser to remove.

If the last top-level definition (before the implicit main)
is a comment, currently need to have a blank line after it.