Get text selection mode working
This commit also comments out all the old code from DeuceWidgets. This
ensures that none of it interferes with the behavior of the new code
(particularly in handling when the Deuce widgets are enabled and when
they are selectable).

All of InterfaceView3 has also been commented out to accommodate this
change, as well as some parts of InterfaceController.