Handle after target polygon in ELet pattern better
The after target polygon for the ELet pattern now renders only one of
the two targets; if the whitespace has been altered, it shows the
manually placed target, otherwise, it shows the default target.

Because of this, the handles with two circles now look good again, so
the number of handles for each Deuce polygon has now been set to two.