Add enableTextEdits flag + functionality
To do this, this commit introduces the Updatable module, which provides
a nice API for variables that require additional Controller.update
commands to be successfully updated.

The motivating example for this is the enableTextEdits flag. Any time
this value is changed, a message has to be sent via a port to JS to
tell the editor to update its readOnly attribute.