Fix phantom dragging bug
Previously, if a user clicked on a draggable element (such as a popup
panel or the resizer) but did not drag, they could then drag elsewhere
in the program and the original draggable element would receive the
delta mouse position drags. Now, drag events are cleared when the mouse
is lifted, so this problem is resolved.