Ability to select features and interpret them as a list of EIds in the program.
Each interpretation is separately fed into Deuce Tools (for now). All synthesis results are displayed.

Probably going to make a special tool set for output selection: the deuce tools do not seem particularly useful.

Still no shape or blob interpretation.

To avoid the combinatorial explosion of interpretations (e.g. 2 interpretations for feature1 times 2 interpretations for feature2 times 2 interpretations for feature3 etc.), as a heuristic, only two interpretations for a feature set are considered: the interpretation farthest back in the program (most “distal”, usually constants), and the interpretation farthest forward in a program (most proximal).

Interpretation of points from x+y feature selection is “wrong”. Namely, the `parents` field on a value is sufficient for finding container values that hold the value of interest, but are not “provenance” in that the container value might be on a different execution path: in such a case changing the container could not possible affect the selected feature value. Need to revisit.

The val basedOn-tree provenance still makes sense, however.