lazily run (and cache) Deuce tool results
Before, all Deuce tool results (i.e. Exps) were run by the view. Now,
a result is run only when the result is hovered (in the second-level
menu) for the first time. The resulting Preview is cached for next
time it's hovered. This cache also records CSS class names to handle
cases in which the result of evaluation does not match the expected
safety.