Make evaluation syntax-aware
The evaluator will now emit error messages in the correct syntax. The
only exception to this is for the unparsing of "base values," whose
functions are now isolated in `ValUnparser.elm`. I think this file
should eventually be separated into separate Little unparsing and Elm
unparsing (as the unparsers are for everything else), but I'm not sure.
It's not such a big deal for now, though, because the biggest difference
in terms of base values between Little and Elm is just the presence
of commas in lists.