preludeLeo.elm
starting with automatically unparsed version,
stripped out a huge section of SVG-specific functions from the middle,
manually fixed parsing errors.
this file parses when pasted in to the app, but now when i tried
compiling a new top-level prelude parsed by (and inside of) ElmParser;
this failed attempt isn't checked in anywhere yet.