new NewModelAndCmd Msg; changeDomValue from SVG widgets
NewModelAndCmd (Model -> (Model, Cmd Msg)) allows a command to be issued
in addition to udpating the Model. eventually, this probably replace
the existing Msg, so that issueCommand(BasedOnCaption) is structured
better.

changeDomValue is starting to tell the DOM about changes through shape
widgets (right now, only ZInterior for "rect"; and the accompanying
zones are not being updated).