Clean up Makefile
`make clean` now no longer builds, but it deletes more build output that
is sometimes useful to clean up in order to get a truly fresh build.
`make` now does what `make clean` used to do, which I think is more
intuitive. A TODO is to have all files that are in build/out be copied
from elsewhere - currently, some of those files are checked in to the
repo so `make clean` can't just do `rm -rf build/out`