revive hover subscriptions
onMouseOver and onMouseLeave events were being listened to by
polygons and being dispatched to controllers, implying that the
controller was actually using them to update the UI. But the logic
that allowed a change of hover state to affect the UI had been
deleted, rendering these listeners useless. This commit revives
the necessary "subscriptions" so that a change to the hovered
widgets actually refreshes the UI and therefore invoking the
mouse-over and mouse-leave messages actually has an effect.