Make UI less glacial
Was doing too much program traversal for lots of label widgets that don’t even appear in the output.

But now have to include what selectable item is associated with the pat being renamed so we can fast exit without traversing the whole provenance/program to ensure the label in question does not correspond to the pattern being editted.

Also incidentally fixes the UI un-niceness that all copies of a pat in the output would appear and be edited simultaneously when editing one.