Clear keys when hitting “Run” so that it can clear that UI bug of keys getting “stuck” down in the model
Cmd-tab out can have the cmd key getting stuck “down” in the model, leading to various UI bugs.

Since hitting “Run” again is a common response to UI bugs, we’ll have it clear the keysDown.

Solving the root cause is harder—we’re already trying to clear the keysDown on visibility changes but it doesn’t always seem to work.