childCodeObjects now returns targets as siblings of adjacent elements
Previously, `childCodeObjects` would return targets before and after an
element as children of that element. This made it difficult to get Deuce
selection/hover polygons working correctly. This change makes it so that
a before target is instead the prior sibling of the element it precedes.
After targets are removed in the general case since they're not always
appropriate, but they're added in several cases, such as after list
elements, where they are useful but previously missing.
This commit also makes other changes, so that text selection should
still work, and so that the `Exp` comments are more accurate