Types are now tagged with a set of semantic roles. Type alias kick off the tagging. The tags are propagated during unification. This:
(a) Fixes the prelude function tools so they show up again, and
(b) Allows those semantic labels to propogate to user-defined functions in the program (so drawing said functions results in saner defaults)

Since only type vars are looked up in the equivalenceClasses data structure that keeps track of substitutions, there may be some cases where the roles are not propagated correctly (unification between two non-typevars). The fix is probably to introduce a type var first in that case and then unify…and make sure every branch of AlgorithmJish.infer returns a recursive call to infer or a type var. We’ll fix as needed.