revert the change to mapFoldExp in cd25da00
cd25da00 changed mapFoldExp to a cleaner simpler version - however, this
change is the source of bugs. Namely, if a program consists entirely of
`( (3) )`, it is impossible to use deuce to select the outer
parentheses. This is only one of several bugs I have noticed regarding
parentheses. I suspect that the new version of mapFoldExp is somehow
failing to do a necessary step for the parent, after it has done the
appropriate work on the children, and that this is messing up the ids
(or some other facet) for `EParen`s. However, I don't understand
mapFoldExps nor childCodeExtractors well enough to figure out exactly
what the problem is. So for now I have to revert it, and it can be
revived once a solution is found