TSEFLLP: Get lists to work somewhat reasonably.
Because “List” in Leo denotes a TList, we have to add the List definition to our datatype environment manually. Writing the equivalent definition in Leo is not parsed like all other datatypes.

Also, need a type annotation on the value of interest so we can instantiate type variables (List a -> List Num).